Support Customize CSS

+++ src/Moonglade.Web/Controllers/SettingsController.cs
@@ -26,6 +26,7 @@ using Moonglade.Setup;
 using Moonglade.Web.Filters;
 using Moonglade.Web.Models;
 using Moonglade.Web.Models.Settings;
+using NUglify;
 using X.PagedList;
 
 namespace Moonglade.Web.Controllers
@@ -613,10 +614,7 @@ namespace Moonglade.Web.Controllers
         [HttpPost("security")]
         public async Task<IActionResult> Security(SecuritySettingsViewModel model)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest();
-            }
+            if (!ModelState.IsValid) return BadRequest();
 
             var settings = _blogConfig.SecuritySettings;
             settings.WarnExternalLink = model.WarnExternalLink;
@@ -628,6 +626,56 @@ namespace Moonglade.Web.Controllers
             _blogConfig.RequireRefresh();
 
             await _blogAudit.AddAuditEntry(EventType.Settings, AuditEventId.SettingsSavedAdvanced, "Security Settings updated.");
+            return Ok();
+        }
+
+        #endregion
+
+        #region CustomCss
+
+        [HttpGet("custom-css")]
+        public IActionResult CustomStyleSheet()
+        {
+            var settings = _blogConfig.CustomStyleSheetSettings;
+            var vm = new CustomStyleSheetSettingsViewModel
+            {
+                EnableCustomCss = settings.EnableCustomCss,
+                CssCode = settings.CssCode
+            };
+
+            return View(vm);
+        }
+
+        [HttpPost("custom-css")]
+        public async Task<IActionResult> CustomStyleSheet(CustomStyleSheetSettingsViewModel model)
+        {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
+            var settings = _blogConfig.CustomStyleSheetSettings;
+
+            if (model.EnableCustomCss && string.IsNullOrWhiteSpace(model.CssCode))
+            {
+                ModelState.AddModelError(nameof(CustomStyleSheetSettingsViewModel.CssCode), "CSS Code is required");
+                return BadRequest(ModelState);
+            }
+
+            var uglifyTest = Uglify.Css(model.CssCode);
+            if (uglifyTest.HasErrors)
+            {
+                foreach (var err in uglifyTest.Errors)
+                {
+                    ModelState.AddModelError(model.CssCode, err.ToString());
+                }
+                return BadRequest(ModelState);
+            }
+
+            settings.EnableCustomCss = model.EnableCustomCss;
+            settings.CssCode = model.CssCode;
+
+            await _blogConfig.SaveConfigurationAsync(settings);
+            _blogConfig.RequireRefresh();
+
+            await _blogAudit.AddAuditEntry(EventType.Settings, AuditEventId.SettingsSavedAdvanced, "Custom Style Sheet Settings updated.");
             return Ok();
         }
 
@@ -723,10 +771,7 @@ namespace Moonglade.Web.Controllers
         {
             static void DeleteIfExists(string path)
             {
-                if (Directory.Exists(path))
-                {
-                    Directory.Delete(path);
-                }
+                if (Directory.Exists(path)) Directory.Delete(path);
             }
 
             try
@@ -790,7 +835,7 @@ namespace Moonglade.Web.Controllers
                 ModelState.AddModelError("username", $"User '{model.Username}' already exist.");
                 return Conflict(ModelState);
             }
-            
+
             var uid = await accountService.CreateAsync(model.Username, model.Password);
             return Json(uid);
         }
